% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{categorical_sample_size}
\alias{categorical_sample_size}
\title{Estimate the sample size for a variable in function of its categories.}
\usage{
categorical_sample_size(categories, increment = 10)
}
\arguments{
\item{categories}{A vector containing the number of categories of
each variable.}

\item{increment}{A number as a constant to which the sample size is
incremented as a product.}
}
\value{
The sample size for a categorical variable based on a
    ordered permutation heuristic approximation of its categories.
}
\description{
Estimate the sample size for a variable in function of its categories.
}
