% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_informative_variable}
\alias{new_informative_variable}
\title{Create an informative uniform categorical random variable.}
\usage{
new_informative_variable(variable_labels, variable_class,
  information_level = 1)
}
\arguments{
\item{variable_labels}{A factor as the labels for the new
informative variable.}

\item{variable_class}{A factor as the class of the variable.}

\item{information_level}{A integer as the information level of the
new variable.}
}
\value{
A factor that represents an informative uniform categorical
    random variable created using the Kononenko method.
}
\description{
The sampling for the items of the created variable is done with replacement.
}
