% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_xor_variables}
\alias{new_xor_variables}
\title{Create a set of categorical variables using the logical XOR operator.}
\usage{
new_xor_variables(n_variables = 2, n_instances = 1000, noise = 0)
}
\arguments{
\item{n_variables}{An integer as the number of variables to be
created. It is the number of column variables of the table, an
additional column is added as a result of the XOR operator over
the instances.}

\item{n_instances}{An integer as the number of instances to be
created. It is the number of rows of the table.}

\item{noise}{A float number as the noise level for the variables.}
}
\value{
A set of random variables constructed using the logical XOR
    operator.
}
\description{
Create a set of categorical variables using the logical XOR operator.
}
\examples{
new_xor_variables(2, 4, 0)
new_xor_variables(5, 10, 0.5)
}
