% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rel_freq}
\alias{rel_freq}
\title{Relative frequency of values of a categorical variable.}
\usage{
rel_freq(variable)
}
\arguments{
\item{variable}{A factor as a categorical variable}
}
\value{
Relative frecuency distribution table for the values in
    \code{variable}.
}
\description{
Relative frequency of values of a categorical variable.
}
\examples{
rel_freq(factor(c(0,1)))
rel_freq(factor(c('a','a','b')))
\dontrun{
rel_freq(c(0,1))
}
}
