% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_size}
\alias{sample_size}
\title{Estimate the sample size for a categorical variable.}
\usage{
sample_size(max, min = 1, z = 1.96, error = 0.05)
}
\arguments{
\item{max}{A number as the maximum value of the possible
categories.}

\item{min}{A number as the minimum value of the possible
categories.}

\item{z}{A number as the confidence coefficient.}

\item{error}{Admissible sampling error.}
}
\value{
The sample size for a categorical variable based on a
    variance heuristic approximation.
}
\description{
Estimate the sample size for a categorical variable.
}
