% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\alias{shannon_entropy}
\alias{H}
\title{Estimation of Shannon entropy for a categorical variable.}
\usage{
shannon_entropy(x)

H(x)
}
\arguments{
\item{x}{A factor as the represented categorical variable.}
}
\value{
Shannon entropy estimation of the categorical variable.
}
\description{
The Shannon entropy estimates the average minimum number of bits
needed to encode a string of symbols, based on the frequency of the
symbols (see
\url{http://www.bearcave.com/misl/misl_tech/wavelets/compression/shannon.html}).
}
\examples{
shannon_entropy(factor(c(1,0)))
shannon_entropy(factor(c('a','b','c')))
\dontrun{
shannon_entropy(1)
shannon_entropy(c('a','b','c'))
}
}
